+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- feat: idempotency for tasks (#77)
+
 ## [1.0.0-rc.7] - 2026-04-07
 
 - bump: to v1.0.0-rc.7 (#68)